% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfusionMatrix_and_otherMetrics.R
\name{confusion_metrics}
\alias{confusion_metrics}
\title{Generate Confusion Matrix and Metrics}
\usage{
confusion_metrics(y_true, y_pred, cutoff = 0.5)
}
\arguments{
\item{y_true}{Actual response values (binary: 0 or 1)}

\item{y_pred}{Predicted probabilities}

\item{cutoff}{Cutoff threshold for classification (default is 0.5)}
}
\value{
A list containing confusion matrix and metrics
}
\description{
Computes confusion matrix and diagnostic metrics.
}
